+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add a new component on the map entity, `TiledMapImageRepeatMargin`, to control the margin to apply for repeating images
 - Add a new component, `TiledMapReference`, to track from which map a Tiled item belongs to
+- Add helpers on `TiledMapAsset` to ease the usage of `TiledObject` functions
 
 ### Documentation